🧵 refactor: modularize resolutions-rs for better code organization
Split functionality from module_factory.rs into separate modules:
- module_resolution_error.rs: Error handling types
- process_module.rs: Module processing logic
- read_file.rs: File reading utilities

This improves code organization while maintaining the async and concurrent processing improvements.
