+++ bundler-packages/resolutions-rs/src/lib.rs
@@ -1,10 +1,16 @@
 pub mod models;
 pub mod module_factory;
+pub mod module_resolution_error;
+pub mod process_module;
+pub mod read_file;
 pub mod resolve_import_path;
 pub mod resolve_imports;
 
 // Re-export the main functions and types for easier access
 pub use models::{ModuleInfo, ResolvedImport};
 pub use module_factory::module_factory;
+pub use module_resolution_error::ModuleResolutionError;
+pub use process_module::process_module;
+pub use read_file::read_file;
 pub use resolve_import_path::resolve_import_path;
 pub use resolve_imports::resolve_imports;
